LangChain/requirements.txt
langchain==0.2.6
langchain-community==0.2.6
python-dotenv==1.0.1
mlflow==2.14.2
